rename metadata.go to stream.go
