--- metadata.go
@@ -1,49 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-package winmd
-
-import (
-	"io"
-)
-
-// Stream provides access to metadata tables and heaps,
-// as defined in §II.24.
-type Stream struct {
-	StreamHeader
-
-	// Embed ReaderAt for ReadAt method.
-	// Do not embed SectionReader directly
-	// to avoid having Read and Seek.
-	// If a client wants Read and Seek it must use
-	// Open() to avoid fighting over the seek offset
-	// with other clients.
-	io.ReaderAt
-	sr *io.SectionReader
-}
-
-// Data reads and returns the contents of the stream s.
-func (s *Stream) Data() ([]byte, error) {
-	dat := make([]byte, s.sr.Size())
-	n, err := s.sr.ReadAt(dat, 0)
-	if n == len(dat) {
-		err = nil
-	}
-	return dat[0:n], err
-}
-
-// Open returns a new ReadSeeker reading the stream s.
-func (s *Stream) Open() io.ReadSeeker {
-	return io.NewSectionReader(s.sr, 0, 1<<63-1)
-}
-
-// Stream returns the stream with the given name, or nil if no such
-// stream exists.
-func (m *File) Stream(name string) *Stream {
-	for _, s := range m.Streams {
-		if s.Name == name {
-			return s
-		}
-	}
-	return nil
-}
